CDR-2499 introduced StoredQueryDefinition
