+++ generator-commons/src/main/java/org/ehrbase/openehr/sdk/generator/commons/aql/parameter/StoredQueryDefinition.java
@@ -1,0 +1,56 @@
+/*
+ * Copyright (c) 2022 vitasystems GmbH and Hannover Medical School.
+ *
+ * This file is part of project openEHR_SDK
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.ehrbase.openehr.sdk.generator.commons.aql.parameter;
+
+import java.util.Optional;
+import org.apache.commons.lang3.StringUtils;
+
+public class StoredQueryDefinition {
+
+    private final String qualifiedQueryName;
+    private final String version;
+    private String type;
+
+    public StoredQueryDefinition(String qualifiedQueryName, String version) {
+        this.qualifiedQueryName = qualifiedQueryName;
+        this.version = version;
+    }
+
+    public StoredQueryDefinition(String qualifiedQueryName, String version, String type) {
+        this.qualifiedQueryName = qualifiedQueryName;
+        this.version = version;
+        this.type = type;
+    }
+
+    public StoredQueryDefinition type(String value) {
+        this.type = value;
+        return this;
+    }
+
+    public Optional<String> getType() {
+        return Optional.ofNullable(this.type);
+    }
+
+    public boolean isValid() {
+        return !StringUtils.isEmpty(qualifiedQueryName) && !StringUtils.isEmpty(version);
+    }
+
+    public String getPath() {
+        return qualifiedQueryName + "/" + version;
+    }
+}
